feat: Add sides points rotations for specific stones Specific stones are : - echo fighters, - Cloud, Corrin and Bayonetta (because of the Player 2 figures), - Pyra and Mythra (who have mirrored rotation), - Steve and Alex (who have mirrored rotation too)
Also, I optimise every forEach in need by throwing a BreakException
so it doesn't slow the loading when it already found the desired value.

And finally, a new field has appeared in every object:
`switchIndex`.
For now it is only usefull for Pyra and Mythra to change easily but it
can be use in the future (Pokemon trainer ?).